Refactor `RewriteBootstrapScriptsStepSpecs` tests
